+++ simworld.h
@@ -657,7 +657,7 @@ public:
 	 */
 	uint32 speed_to_tiles_per_month(uint32 speed) const
 	{
-		const int left_shift = ticks_per_world_month_shift - YARDS_PER_TILE_SHIFT;
+		const int left_shift = (int)(ticks_per_world_month_shift - YARDS_PER_TILE_SHIFT);
 		if (left_shift >= 0) {
 			return speed << left_shift;
 		} else {
@@ -967,7 +967,7 @@ public:
 	inline bool ist_markiert(const grund_t* gr) const { return marker.ist_markiert(gr); }
 
 	// Getter/setter methods for maintaining the industry density
-	inline uint32 get_target_industry_density() const { return (finance_history_month[0][WORLD_CITICENS] * industry_density_proportion) / 10000; }
+	inline uint32 get_target_industry_density() const { return ((uint32)finance_history_month[0][WORLD_CITICENS] * industry_density_proportion) / 10000; }
 	inline uint32 get_actual_industry_density() const { return actual_industry_density; }
 	
 	inline void decrease_actual_industry_density(uint32 value) { actual_industry_density -= value; }
